Delete resources directory
Deleted the resources directory because I don't think I'll ever use
either of the libraries in it. They were both libraries to try to
interact with the console, but there didn't even seem to be a hacky way
to get arrow keys to work with the console. I would have had to make my
own custom JVM. I thought of using an invisible gui to catch keypresses,
but the invisible gui window would have needed focus, and there doesn't
seem to be a "right" way of doing that. Using an actual gui window,
disguised to look like a console, would be the best option