Fix screen not getting rid of parts that are not there
I finally got the snake to stop making a trace of where it went. Turns
out that the problem was that the cases that controlled which direction
the snake was moving in would move the whole snake in the same
direction, instead of each segment moving into the place where the
previous segment was. I had to make 2 new variables that would store the
x and y of the last last position, and the loops alternate between using
the original prevX prevY and the new prevNextX prevNextY. I know I am
not supposed to, but I did another change, and that was to move the
starting position of the snake to the middle instead of the edge,
because it was deleting part of the border.