Add detection for going backwards
I made it so that the snake cannot go directly back into itself. It can
still intersect with itself though. I did this by adding a second
variable to keep track of direction, but I only change it after I get
through the cases. The older dir variable gets set to tempDir if the
snake tries to go back into itself. One issue though, is that the user
can just hold down the arrow key in the opposite direction it is moving
to stop it.