Add ability to retry if you lose or win
I had make a main loop and I moved a bunch of stuff into it. The main
thing is the part that clears the screen and then adds it back in. I
tried using replaceRange in a clearScreen method, but that just ended up
being a headache. I could never get the range right for whatever reason.
I also added a loop that just waits a specific amount of ticks so that
it can detect keypresses to restart